requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
qdrant-client
chromadb
langchain-openai
langchain-huggingface
sentence-transformers
pytest
pytest-asyncio
pytest-cov
httpx
python-dotenv
python-multipart
pypdf
pandas
openpyxl
pinecone-client
